assets and yml
