+++ Resources/Prototypes/Entities/Structures/Wallmounts/elevator.yml
@@ -1,0 +1,44 @@
+- type: entity
+  id: ElevatorControl
+  name: elevator control box
+  description: The box that plays the cool music.
+  placement:
+    mode: SnapgridCenter
+    snap:
+    - Wallmount
+  components:
+  - type: ElevatorEntrance
+    elevatorMapKey: lavaland
+    entranceId: salvage
+    exitId: salvage
+  - type: ElevatorExit
+    exitId: salvage
+  - type: ElevatorCollision
+    collisionId: elevator
+  - type: DeviceLinkSink
+    ports:
+      - ElevatorFloorChange
+  - type: Physics
+    bodyType: Dynamic
+  - type: Fixtures
+    fixtures:
+      elevator:
+        shape:
+          !type:PhysShapeAabb
+          bounds: "-1.5,-1.5,1.5,1.5"
+        mask:
+        - ItemMask
+        layer:
+        - SlipLayer
+        hard: false
+  - type: Sprite
+    drawdepth: WallMountedItems
+    sprite: Structures/Wallmounts/elevator_control.rsi/
+    state: box
+  - type: InteractionOutline
+  - type: Clickable
+  - type: Transform
+  - type: WallMount
+  - type: Tag
+    tags:
+      - Structure
