+++ Resources/Textures/Structures/Wallmounts/elevator_control.rsi/meta.json
@@ -1,0 +1,14 @@
+{
+    "version": 1,
+    "license": "CC-BY-SA-3.0",
+    "copyright": "Doctor-Cpu",
+    "size":
+    {
+        "x": 32,
+        "y": 32
+    },
+    "states":
+    [{
+        "name": "box"
+    }]
+}
